requirements.txt
opencv-python
mediapipe
numpy
Pillow
torch
torchvision
playsound
pandas
matplotlib
scikit-image
jupyterlab
tqdm
